refactor: backend structure
