+++ backend/service/index.js
@@ -1,0 +1,7 @@
+const getContacts = require('./getContacts');
+const getContactDetail = require('./getContactDetail');
+
+module.exports = {
+  getContacts,
+  getContactDetail,
+};
